requirements.txt
pipasgiref==3.6.0
beautifulsoup4==4.12.2
canvas==0.3
contourpy==1.0.7
crispy==0.7.4
crispy-bootstrap5==0.7
crysp==1.2
cycler==0.11.0
Django==4.2
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-fontawesome-6==1.0.0.0
fabio==2023.4.1
fonttools==4.39.3
freetype-py==2.3.0
h5py==3.8.0
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pisa==3.0.33
psycopg2-binary==2.9.6
pycairo==1.23.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
python-environ==0.4.54
reportlab==4.0.0
rlPyCairo==0.2.0
silx==1.1.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
zipp==3.15.0
